Revert "MainMod: allow fractional kick times"
This reverts commit 088ff6799469415c6b25274007738f6e502166c4.

+++ Cryptex/main_controller.cpp
@@ -39,7 +39,7 @@ void main_controller::kick()
 	controller->send("k");
 }
 
-void main_controller::kick(const double &ms)
+void main_controller::kick(const int &ms)
 {
 	assume_coilgun();
 	controller->send("k", ms);
